server/requirements.txt
grpcio==1.59.0
keras==2.13.1
protobuf==4.24.4
tensorflow==2.13.0
yake==0.4.8
